Refactored validation of kudo.NewClient
